refactor: move `mod tokio_util` to runtime (#12332)
This avoids a bunch of duplicated code.
