+++ cli/lsp/cache.rs
@@ -5,12 +5,12 @@ use crate::cache::FetchCacher;
 use crate::flags::Flags;
 use crate::proc_state::ProcState;
 use crate::resolver::ImportMapResolver;
-use crate::tokio_util::create_basic_runtime;
 
 use deno_core::error::anyhow;
 use deno_core::error::AnyError;
 use deno_core::ModuleSpecifier;
 use deno_runtime::permissions::Permissions;
+use deno_runtime::tokio_util::create_basic_runtime;
 use import_map::ImportMap;
 use std::path::PathBuf;
 use std::thread;
