+++ cli/lsp/diagnostics.rs
@@ -7,13 +7,13 @@ use super::language_server;
 use super::tsc;
 
 use crate::diagnostics;
-use crate::tokio_util::create_basic_runtime;
 
 use deno_core::error::anyhow;
 use deno_core::error::AnyError;
 use deno_core::resolve_url;
 use deno_core::serde_json::json;
 use deno_core::ModuleSpecifier;
+use deno_runtime::tokio_util::create_basic_runtime;
 use log::error;
 use lspower::lsp;
 use std::collections::HashMap;
